Big Update (9/26)
Long time no see. In this update, I bring a new way to building palette part of spnet. That because the CNN palette will learn a fix pixel is fix color, however, that not what we want. We need a palette network could put different color in each pixel position with different input image. The old version CNN palette will let segmentation part in spnet learn to segment each color position, not different class. So we change to use image process to figure this problem. In spnet.py file, we can notice here has two way to get color. One is use torch.mode to get the mode value in each color channel from input image. The orther is use mean. The reason why we choose the second way to get color is the first way will get 0 is the mode value mostly. But that might just some noise (which should be another classes but here seg did not do a good jod).